Generate new DKIM data only if keys do not exist yet

+++ entrypoint.sh
@@ -207,34 +207,37 @@ function start_postfix {
     
     sed -n -e '/^Domain\s/!p' -e '$aDomain '$HOSTNAME -i /etc/opendkim/opendkim.conf
     # DKIM for all virtual domains and $HOSTNAME 
-     if [ "$SMF_DKIM_ALL" != "" ]; then
-        mkdir -p /var/db/dkim/$HOSTNAME
-        cp /var/db/dkim/default.* /var/db/dkim/$HOSTNAME
-        
-        echo "default._domainkey.${HOSTNAME} ${HOSTNAME}:default:/var/db/dkim/${HOSTNAME}/default.private
-" >> /etc/opendkim/KeyTable
+    if [ "$SMF_DKIM_ALL" != "" ]; then
+        if [ ! -f /var/db/dkim/$HOSTNAME/default.private ]; then
+            mkdir -p /var/db/dkim/$HOSTNAME
+            cp /var/db/dkim/default.* /var/db/dkim/$HOSTNAME
 
-        echo "${HOSTNAME} default._domainkey.${HOSTNAME}" >> /etc/opendkim/SigningTable
+            echo "default._domainkey.${HOSTNAME} ${HOSTNAME}:default:/var/db/dkim/${HOSTNAME}/default.private
+            " >> /etc/opendkim/KeyTable
 
-        echo "${HOSTNAME}" >> /etc/opendkim/TrustedHosts
+            echo "${HOSTNAME} default._domainkey.${HOSTNAME}" >> /etc/opendkim/SigningTable
 
+            echo "${HOSTNAME}" >> /etc/opendkim/TrustedHosts
+        fi
         for virtualDomain in $virtualDomains; do
-            mkdir -p /var/db/dkim/${virtualDomain}
-            echo "OpenDKIM: Keys for ${virtualDomain} not found, generating..."
-            opendkim-genkey -b 2048 -d ${virtualDomain} -D /var/db/dkim/${virtualDomain} -s default -v
+            if [ ! -f /var/db/dkim/${virtualDomain}/default.private ]; then
+                mkdir -p /var/db/dkim/${virtualDomain}
+                echo "OpenDKIM: Keys for ${virtualDomain} not found, generating..."
+                opendkim-genkey -b 2048 -d ${virtualDomain} -D /var/db/dkim/${virtualDomain} -s default -v
 
-            chmod 400 /var/db/dkim/${virtualDomain}/default.private
-            chown opendkim:opendkim /var/db/dkim/${virtualDomain}/default.private
+                chmod 400 /var/db/dkim/${virtualDomain}/default.private
+                chown opendkim:opendkim /var/db/dkim/${virtualDomain}/default.private
 
-            echo "default._domainkey.${virtualDomain} ${virtualDomain}:default:/var/db/dkim/${virtualDomain}/default.private
-" >> /etc/opendkim/KeyTable
+                echo "default._domainkey.${virtualDomain} ${virtualDomain}:default:/var/db/dkim/${virtualDomain}/default.private
+            " >> /etc/opendkim/KeyTable
 
-            echo "${virtualDomain} default._domainkey.${virtualDomain}" >> /etc/opendkim/SigningTable
+                echo "${virtualDomain} default._domainkey.${virtualDomain}" >> /etc/opendkim/SigningTable
 
-            echo "${virtualDomain}" >> /etc/opendkim/TrustedHosts
+                echo "${virtualDomain}" >> /etc/opendkim/TrustedHosts
 
-            echo "OpenDKIM: Add TXT record to DNS for ${virtualDomain}:"
-            cat /var/db/dkim/${virtualDomain}/default.txt  
+                echo "OpenDKIM: Add TXT record to DNS for ${virtualDomain}:"
+                cat /var/db/dkim/${virtualDomain}/default.txt  
+            fi
         done
     fi
